+++ CHANGELOG.md
@@ -6,14 +6,17 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- `operator-rs` `0.4.0` → `0.9.0` ([#81]).
+- BREAKING: `ClusterRef` namespace now optional ([#95]).
+- BREAKING: In case the namespace is omitted, the operator defaults to the `TrinoCluster` namespace instead of `default` ([#95]).
 - User authentication now provided via secret instead of custom resource ([#81]).
 - User authentication not exposed in configmap anymore ([#81]).
 - TLS certificates / keystore now retrieved via secret-operator ([#81]).
-- The Trino version is now string instead of enum ([#81]).
+- The Trino version is now a string instead of enum ([#81]).
+- `operator-rs` `0.4.0` → `0.9.0` ([#81], [#95]).
+- `stackable-regorule-crd` `0.2.0` → `0.5.0` ([#81]).
 
 [#81]: https://github.com/stackabletech/trino-operator/pull/81
-
+[#95]: https://github.com/stackabletech/trino-operator/pull/95
 
 ## [0.2.0] - 2021-12-06